Fix codeclimate
